Fix on linux

+++ db/db_impl/db_impl.cc
@@ -718,7 +718,7 @@ void DBImpl::PersistStats() {
   }
   uint64_t now_seconds = env_->NowMicros() / kMicrosInSecond;
 
-  fprintf(stdout, "PST: seconds %llu\n", now_seconds);
+  fprintf(stdout, "PST: seconds %llu\n",  (unsigned long long)now_seconds);
   Statistics* statistics = immutable_db_options_.statistics.get();
   if (!statistics) {
     return;
@@ -810,7 +810,7 @@ void DBImpl::PersistStats() {
                    stats_history_size, stats_history_.size());
   }
   TEST_SYNC_POINT("DBImpl::PersistStats:End");
-  fprintf(stdout, "PST: done seconds %llu\n", now_seconds);
+  fprintf(stdout, "PST: done seconds %llu\n",  (unsigned long long) now_seconds);
 #endif  // !ROCKSDB_LITE
 }
